rename : EventStatusTag => EventTag
