+++ src/components/common/MyEventInfo/MyEventInfo.tsx
@@ -5,7 +5,7 @@ import { Event } from '@/types/event';
 
 import { useNavigate } from 'react-router-dom';
 
-import EventStatusTag from '../EventStateTag/EventStatusTag';
+import EventStatusTag from '../EventTag/EventTag';
 import {
   EventDescription,
   EventInfoSection,
